+++ package.json
@@ -17,5 +17,8 @@
     "app": "lerna exec --scope app -- yarn watch",
     "test": "jest",
     "test:watch": "jest -w"
+  },
+  "dependencies": {
+    "@bufferapp/ui": "^5.60.1"
   }
 }
