+++ packages/app-shell/src/PlanSelector/index.jsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { MODALS } from '../hooks/useModal';
 import { UserContext } from '../context/User';
 import { ModalContext } from '../context/Modal';
+import { PlanSelectorContainer } from './components/PlanSelectorContainer';
 
 const PlanSelector = () => {
   return (
@@ -10,8 +11,8 @@ const PlanSelector = () => {
       {(user) => (
         <ModalContext.Consumer>
           {(modal) => (
-            <PlanSelector
-              planOptions={user}
+            <PlanSelectorContainer
+              planOptions={user.currentOrganization.billing.changePlanOptions}
               openPaymentMethod={() => {
                 modal.openModal(MODALS.paymentMethod);
               }}
